Mostly works with rating
migration, model and controller for the ratings system. Ratings are mostly done with few exceptions (average ratings on movies and logged in users rated movies need to be implemented). Form with auth protection is in place on the movies show view but there's no functionality yet. Problem with admins dashboard page still presists